requirements.txt
tensorflow==2.12.0
keras==2.12.0
flask==2.3.2
opencv-python==4.8.0.74
Pillow==9.5.0
numpy==1.24.3
scikit-learn==1.3.0
werkzeug==2.3.6
matplotlib==3.7.1
gunicorn==21.2.0
